requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
websocket-client>=1.6.0
altair>=5.0.0
